Add files via upload
I am following the Data Science specialization from John Hopkins University on Coursera. since the course is pretty old, and the code examples for the third course, Getting and Cleaning Data are only given in the videos and non-text pdf handouts, I decided to assemble a rmd file with the course code, with link and code updates / workarounds as I found them in the forum. This file only contains the course examples, no test material! So far only week 1, but I will update the file as I finish the next week